requirements.txt


python-dotenv

#gutenbergpy

pandas

plotly


openai # 1.3.8
langchain # 0.0.348 ... 0.0.353
tiktoken
faiss-cpu

# llama:
torch # 2.1.0+cu121 (for colab)
transformers # 4.35.2
accelerate # 0.25.0
# torchtext # 0.16.0



pytest
